requirements.txt
fastapi
uvicorn
requests
faiss-cpu
scikit-learn
